+++ lib/src/progress/progress_bar.dart
@@ -2,9 +2,11 @@ import 'package:flutter/material.dart';
 import 'package:video_player/video_player.dart';
 import 'package:video_player_controls/src/buttons/fast_foward_button.dart';
 import 'package:video_player_controls/src/buttons/fast_rewind_button.dart';
+import 'package:video_player_controls/src/buttons/full_screen_button.dart';
 import 'package:video_player_controls/src/buttons/next_button.dart';
 import 'package:video_player_controls/src/buttons/play_button.dart';
 import 'package:video_player_controls/src/buttons/previous_button.dart';
+import 'package:video_player_controls/src/buttons/subtitle_button.dart';
 import 'package:video_player_controls/src/progress/progres_slider.dart';
 import 'package:video_player_controls/src/progress/video_period.dart';
 import 'package:video_player_controls/video_player_controls.dart';
@@ -40,45 +42,55 @@ class _ProgressBarState extends State<ProgressBar> {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
-      padding: const EdgeInsets.all(8.0),
-      child: Container(
-        child: Row(
-          children: <Widget>[
-            Expanded(
-              child: new Container(
-                child: new Column(
-                  children: <Widget>[
-                    VideoPeriod(
-                      videoPlayerController: widget.videoPlayerController,
-                    ),
-                    new ProgressSlider(
-                      duration: duration,
-                    ),
-                    Row(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: <Widget>[
-                        _controller.showSeekButtons == false
-                            ? new Container()
-                            : new FastRewindButton(),
-                        _controller.showSkipButtons == false
-                            ? new Container()
-                            : new PreviousButton(),
-                        new PlayButton(),
-                        _controller.showSkipButtons == false
-                            ? new Container()
-                            : new NextButton(),
-                        _controller.showSeekButtons == false
-                            ? new Container()
-                            : new FastFowardButton(),
-                      ],
-                    ),
-                  ],
-                ),
+    return Container(
+      child: Row(
+        children: <Widget>[
+          Expanded(
+            child: new Container(
+              child: new Column(
+                children: <Widget>[
+                  VideoPeriod(
+                    videoPlayerController: widget.videoPlayerController,
+                  ),
+                  new ProgressSlider(
+                    duration: duration,
+                  ),
+                  Row(
+                    children: <Widget>[
+                      Expanded(
+                        child: Row(
+                          mainAxisAlignment: MainAxisAlignment.center,
+                          children: <Widget>[
+                            _controller.showSeekButtons == false
+                                ? new Container()
+                                : new FastRewindButton(),
+                            _controller.showSkipButtons == false
+                                ? new Container()
+                                : new PreviousButton(),
+                            new PlayButton(),
+                            _controller.showSkipButtons == false
+                                ? new Container()
+                                : new NextButton(),
+                            _controller.showSeekButtons == false
+                                ? new Container()
+                                : new FastFowardButton(),
+                          ],
+                        ),
+                      ),
+                      widget.controller.hasSubtitles == true
+                          ? SubtitleButton()
+                          : new Container(),
+                      widget.controller.fullScreenByDefault == true ||
+                              widget.controller.allowFullScreen == false
+                          ? new Container()
+                          : new FullScreenButton(),
+                    ],
+                  ),
+                ],
               ),
             ),
-          ],
-        ),
+          ),
+        ],
       ),
     );
   }