+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## [2.0.1] - Wed July 15, 2020
+## [2.0.2] - Wed July 16, 2020
 
-* Remove unused imports
+* Correct RangeError(index) exception
+* Add allowScreenSleep functionality
+* Add fullScreenDefault feature
+* Add fullscreen buttons